TP8

+++ src/main/java/fr/diginamic/appliweb/entities/Departement.java
@@ -32,7 +32,7 @@ public class Departement {
     // GETTERS / SETTERS
 
     public void setId(int id) {
-        this.id = id;
+        this.id = (long) id;
     }
 
     public Long getId() {
